Use commands.json as the versioning source
When commands are installed, the VCS info can be pruned by Composer. Further,
there's no guarantee they were installed through Git, which is all that the
default implementation supported. This change removes the version delegation
from the package and puts it back in Plow itself (which was baked out during
command installation as of the previous commit). As a nice side-effect, it
*forces* all version info to stay in sync since there can be only one source of
truth (the version Composer knows about)